requirements.txt
argparse
Babel>=1.3
eventlet>=0.13.0
httplib2>=0.7.5
iso8601>=0.1.9
kombu>=2.4.8
netaddr>=0.7.6
oslo.config>=1.2.0
oslo.messaging>=1.3.0
pbr>=0.6,!=0.7,<1.0
pecan>=0.4.5
six>=1.4.1
WSME>=0.6
